Refactored helpx-internal to have a helpxmain-article template
