+++ templates/helpxmain-article/helpxmain-article.css
@@ -1,0 +1,388 @@
+
+body {
+  line-height: 21px;
+}
+
+body table {
+  font-size: 16px;
+  border-spacing: 0;
+  table-layout: fixed;
+  width: 100%;
+}
+
+body ul li {
+  list-style-type: disc;
+}
+
+table, th, td {
+  padding: 9px 10px;
+  border: 1px solid var(--table-color-border)!important;
+  border-collapse: collapse!important;
+}
+
+table th {
+  background-color: var(--table-header-color-background);
+}
+
+.table-grey table {
+  background: #f7f8f7;
+}
+
+.table-banded-rows tr:nth-child(odd) {
+  background: #f7f8f7;
+}
+/*
+ *  WARNING: This is a temporary fix. It would be better to turn it into a page-description block 
+ *  Look for the .page-description selector on 
+ *  https://helpx-internal.corp.adobe.com/content/help/en/internal/individual/policy-process/vat-rates-direct-store-emea.html
+ */
+h3:first-of-type {
+  color: #666;
+  margin 10px 0 16px;
+  line-height: 1.4;
+  font-size: 18px;
+  font-weight: 500;
+}
+
+li {
+  margin: 6px 0; 
+}
+
+body p {
+  margin-block-start: 0;
+  margin-block-end: 0;
+  font-size: 16px;
+  line-height: 1.5;
+  margin: 10px 0;
+  overflow-wrap: break-word;
+}
+body div {
+  font-size: 16px;
+  line-height: 1.5;
+  overflow-wrap: break-word;
+}
+
+.center {
+  text-align: center;
+}
+
+header.gnav > div {
+  margin: 0;
+}
+
+/* equivalent of section title in AEM */
+.section > .content > h2:not(.first) {
+  font-size: 2rem;
+  font-weight: 400;
+  line-height: 1em;
+  border-top: 1px solid var(--static-color-heading-border);
+  margin-top: 3.5rem;
+  margin-bottom: 0;
+  padding: 20px  0;
+
+}
+
+.section > .content > h2.first {
+  font-size: 2rem;
+  font-weight: 400;
+  line-height: 1em;
+  margin: 0;
+  padding: 20px  0;
+}
+
+.columns {
+  max-width: 1200px;
+  margin: 0 auto;
+}
+
+.accordion dd div {
+  padding-top: 10px;
+}
+
+.accordion ul, .accordion ol {
+  padding-left: 24px;
+}
+
+body.helpxmain-article {
+  font-family: adobe-clean,Source Sans Pro,-apple-system,BlinkMacSystemFont,Segoe UI,Roboto,sans-serif;
+}
+
+body h1 {
+  max-width: 1050px;
+  line-height: 1.15;
+  font-size: 36px;
+  font-weight: 500;
+  text-align: left;
+  word-break: break-word;
+  margin: 0 auto;
+  padding: 0 2.5rem;
+  min-height: 72px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+}
+
+.draft {
+  display: none !important;
+}
+
+/*
+* title
+*/
+
+main .section.dark.xs-spacing.page-title {
+  padding: 0 2.5rem 10px 0 !important;
+  margin: 0;
+  font-size: 36px;
+  font-weight: 500;
+  color: var(--static-color-white)!important;
+  width: 100%;
+  height:auto;
+  z-index: 20;
+  position: fixed;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+}
+
+main .section.dark.xs-spacing.page-title > .content {
+  margin: 0;
+}
+
+/*
+  internal banner
+*/
+
+main .internal-banner .banner {
+  display: flex;
+  gap: 5px;
+  align-items: center;
+  justify-content: center;
+  background-color: var(--internal-banner-color-background);
+  font-size: 14px;
+  font-weight: 700;
+  padding: 12px 0;
+  line-height: 24px;
+  margin: 0;
+  text-transform: uppercase;
+  color: var(--static-color-white);
+}
+
+main .internal-banner {
+  position: absolute;
+  width: 100%;
+  margin: 0;
+}
+/* info icon - added via global block function in scripts.js */
+main .internal-banner .banner .icon {
+  height: 18px;
+  width: 18px;
+}
+
+main .last-updated.content {
+  font-weight: lighter;
+  font-size: 16px;
+  color: var(--internal-banner-color-font);
+  padding: 32px 2.5rem;
+  margin: 0 auto !important;
+  max-width: 1050px;
+}
+
+/*
+* sections
+*/
+
+main > div.section:not(.internal-banner, .page-title) {
+  max-width: 1050px;
+  padding: 0 2.5rem;
+  margin: 0 auto;
+  display: grid;
+  grid-template-columns: 66% 1fr;
+}
+
+main > div.section:not(.internal-banner, .page-title) > div {
+  grid-column-start: 1;
+  grid-column-end: 1;
+}
+
+body > main > .section > .content {
+  margin:0;
+}
+
+
+/*
+ * internal confidential
+ */
+
+.internal-confidential:before {
+  color: #c00;
+  font-weight: 700;
+  content: 'Internal';
+}
+.internal-confidential > div {
+  margin: 4px 0 20px 0;
+  padding: 6px 2px;
+  background: url('https://main--helpx-internal--adobecom.hlx.page/media_1b45f544d71effad595a85747635d01832a3c0149.png') 20px -20px #eaeaea;
+}
+
+/*
+ * footer
+ */
+
+footer {
+  background: var(--footer-color-background);
+  color: var(--footer-color-font);
+  font-size: 11px;
+  margin-top: 20px;
+  padding-top: 20px;
+  padding: 20px 0;
+  position: relative;
+  text-align: center;
+}
+
+/*
+ * on this page
+ */
+
+.on-this-page {
+  position: fixed;
+  max-height: 700px;
+  overflow: auto;
+  width: 20%;
+  top: 214px;
+  right: 40px;
+  display: flex;
+  flex-direction: column;
+}
+
+.on-this-page:before {
+  content: 'Internal';
+  color: #c00;
+  font-weight: 700;
+}
+
+.on-this-page > .content {
+  display: flex;
+  flex-direction: column;
+  gap: 2px;
+  top: 20px;
+  background: url('https://main--helpx-internal--adobecom.hlx.page/media_1b45f544d71effad595a85747635d01832a3c0149.png');
+  background-color: var(--on-this-page-background);
+  padding: 10px;
+  margin: 0;
+}
+
+.on-this-page h5 {
+  margin: 0 0 auto;
+  display: block;
+  margin-inline-start: 0px;
+  margin-inline-end: 0px;
+
+}
+
+.on-this-page a, .on-this-page p {
+  margin: 0 0 0 6px;
+}
+
+a.button:hover, a.button:focus, button:hover, button:focus  {
+  background-color: var(--link-hover-color);
+  cursor: pointer;
+}
+
+button:disabled, button:disabled:hover {
+  background-color: var(--overlay-background-color);
+  cursor: unset;
+}
+
+a.button.secondary, button.secondary {
+  background-color: unset;
+  border: 2px solid currentcolor;
+  color: var(--text-color)
+}
+
+/* phone */
+@media screen and (max-width: 37.4375rem) {
+  .hidden-mobile {
+    display: none !important;
+  }
+
+  body.helpxmain-article h1 {
+    font-size: 1.5rem;
+    padding: 13px 2.5rem;
+  }
+  .layout-container .content-container .section {
+    padding: 0;
+  }
+  main > div.section:not(.internal-banner, .page-title) {
+    padding-left: 2.5rem !important;
+    padding-right: 2.5rem !important;
+  }
+  .page-title + div.section:not(.internal-banner, .page-title), 
+  .internal-banner + div.section:not(.internal-banner, .page-title) {
+    padding-top: 234px;
+  }
+  .last-updated {
+    padding-left: 2.5rem !important;
+  }
+}
+
+/* tablet */
+@media screen and (min-width: 37.5rem) and (max-width: 63.9375rem) {
+  .hidden-tablet {
+    display: none !important;
+  }
+  body.helpxmain-article h1 {
+    font-size: 1.5rem;
+    padding: 13px 2.5rem;
+  }
+  .layout-container .content-container .section {
+    padding: 0;
+  }
+  main > div.section:not(.internal-banner, .page-title) {
+    padding-left: 2.5rem !important;
+    padding-right: 2.5rem !important;
+  }
+  .page-title + div.section:not(.internal-banner, .page-title), 
+  .internal-banner + div.section:not(.internal-banner, .page-title) {
+    padding-top: 233px;
+  }
+  .last-updated {
+    padding-left: 2.5rem !important;
+  }
+  .on-this-page {
+    top: 265px;
+  }
+}
+
+/* desktop */
+@media screen and (min-width: 64rem) {
+  /* mitigate CLS */
+
+  .page-title + div.section:not(.internal-banner, .page-title), 
+  .internal-banner + div.section:not(.internal-banner, .page-title) {
+    padding-top: 202px;
+  }
+
+  .hidden-desktop {
+    display: none !important;
+  }
+
+  .page-title {
+    height: 62px;
+  }
+}
+/* above 1500px */
+@media screen and (min-width: 93.75rem) {
+  body.helpxmain-article h1 {
+    font-size: 2rem;
+  }
+}
+
+@media (max-width: 700px) {
+  .on-this-page {
+    display: none;
+  }
+  main > div.section:not(.internal-banner, .page-title) {
+    grid-template-columns: 100%
+  }
+}
